+++ src/events/EmailEvent.php
@@ -2,6 +2,7 @@
 namespace verbb\workflow\events;
 
 use verbb\workflow\elements\Submission;
+use verbb\workflow\models\Review;
 
 use craft\elements\User;
 use craft\events\CancelableEvent;
@@ -14,5 +15,6 @@ class EmailEvent extends CancelableEvent
     public mixed $mail = null;
     public ?User $user = null;
     public ?Submission $submission = null;
+    public ?Review $review = null;
 
 }