+++ src/events/PrepareEmailEvent.php
@@ -2,6 +2,7 @@
 namespace verbb\workflow\events;
 
 use verbb\workflow\elements\Submission;
+use verbb\workflow\models\Review;
 
 use craft\elements\User;
 use craft\events\CancelableEvent;
@@ -15,5 +16,6 @@ class PrepareEmailEvent extends CancelableEvent
     public ?User $editor = null;
     public ?array $publishers = null;
     public ?array $reviewers = null;
+    public ?Review $review = null;
 
 }